mongodb-inserter/requirements.txt
pymongo==3.10.1
kafka-python==2.0.1
